Refactor code to separate route files
