+++ src/lib.rs
@@ -1,26 +1,3 @@
-use actix_web::{dev::Server, get, post, web, App, HttpResponse, HttpServer, Responder};
-use serde::Deserialize;
-use std::net::TcpListener;
-
-#[get("/health_check")]
-async fn health_check() -> impl Responder {
-    HttpResponse::Ok().finish()
-}
-
-#[derive(Deserialize)]
-struct FormData {
-    email: String,
-    name: String,
-}
-
-#[post("/subscriptions")]
-async fn subscribe(_form: web::Form<FormData>) -> HttpResponse {
-    HttpResponse::Ok().finish()
-}
-
-pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
-    let server = HttpServer::new(|| App::new().service(health_check).service(subscribe))
-        .listen(listener)?
-        .run();
-    Ok(server)
-}
+pub mod configuration;
+pub mod routes;
+pub mod startup;
