+++ src/routes/subscriptions.rs
@@ -1,0 +1,13 @@
+use actix_web::{post, web, HttpResponse};
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+struct FormData {
+    email: String,
+    name: String,
+}
+
+#[post("/subscriptions")]
+pub async fn subscribe(_form: web::Form<FormData>) -> HttpResponse {
+    HttpResponse::Ok().finish()
+}
