+++ src/startup.rs
@@ -1,0 +1,10 @@
+use crate::routes::*;
+use actix_web::{dev::Server, App, HttpServer};
+use std::net::TcpListener;
+
+pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
+    let server = HttpServer::new(|| App::new().service(health_check).service(subscribe))
+        .listen(listener)?
+        .run();
+    Ok(server)
+}
